Add files via upload
Imagine that you have an annual Sales Report based on the financial indicators for the year, broken down by months in which you make specific calculations. This report is downloaded as a standard from a certain database and always has the same column field names, but contains very lack of data so you have to calculate additional data later using formulas. But what if you need to compare not the months of one year, but the years themselves? For example, if you have annual reports for 2020, 2021, and 2022 and need to analyze the indicators of the months by year? 
I will show you several options:
Option 1. 
Additional calculations using Excel formulas in each file,  then combining them using Power Query and creating a visualization using Pivot Tables with Pivot Dashboards. It’s fully 
(Financial Year 2020-2022 are source files, Dashboards (v1) is the completed report)